Lab 3 done.

+++ lab3/task1.c
@@ -6,7 +6,7 @@ int main() {
 	pid_t pid, ppid;
 	pid = getpid();
 	ppid = getppid();
-	printf("My pid = %d\nMy ppid = %d\n", (int) pid, (int) ppid);
+	printf("My pid = %d\nMy ppid = %d\n", pid, ppid);
 	return 0;
 }
